requirements.txt
yfinance~=0.2.37
pandas~=2.0.3
psycopg2~=2.9.9
SQLAlchemy~=2.0.28
dash~=2.16.1
numpy~=1.23.5
requests~=2.31.0
prophet~=1.1.5
sklearn~=0.0.post7
scikit-learn~=1.3.2
statsmodels~=0.14.1
django-googlesearch~=0.2
transformers~=4.30.2
openpyxl~=3.1.2
beautifulsoup4~=4.12.3
lxml~=5.1.0